+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unpublished
 
+- Making `init` ask for workspace and template if none is provided
+
 ## 1.0.2 (2025-01-24)
 
 - Fixing the installation instrucions in the README